Change slack coefficient type from double to long.

+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/core/fees/FeeMarket.java
@@ -29,9 +29,9 @@ public interface FeeMarket {
 
   long getPerTxGaslimit();
 
-  double getSlackCoefficient();
+  long getSlackCoefficient();
 
   static FeeMarket eip1559() {
-    return new FeeMarketConfig(8L, 10000000L, 2.0, 800000L, 10L, 1000000000L, 8000000L);
+    return new FeeMarketConfig(8L, 10000000L, 2L, 800000L, 10L, 1000000000L, 8000000L);
   }
 }
